# docs/swagger.yaml
basePath: /
definitions:
  dto.APIResponse:
    properties:
      data: {}
      message:
        type: string
      request_id:
        type: string
      status:
        type: string
    type: object
...
  dto.CreateUserRequest:
    properties:
      email:
        type: string
      name:
        type: string
      organization_id:
        type: integer
      password:
        type: string
      role:
        type: string
    type: object
...
  dto.ErrorResponse:
    properties:
      message:
        type: string
      request_id:
        type: string
      status:
        type: string
    type: object
...
  dto.UserResponse:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      organization_id:
        type: integer
      role:
        $ref: '#/definitions/model.UserRole'
    type: object
...
  model.UserRole:
    enum:
    - ADMIN
    - DISPATCHER
    - DRIVER
    - ANALYTIC
    type: string
    x-enum-varnames:
    - UserRoleAdmin
    - UserRoleDispatcher
    - UserRoleDriver
    - UserRoleAnalytic
...
host: localhost:8080
info:
  contact: {}
  description: Бэкенд системы мониторинга и телематики автопарка
  title: FleetTrack API
  version: "1.0"
paths:
...
  /users:
    get:
      parameters:
      - description: Фильтр по роли
        in: query
        name: role
        type: string
      - description: Лимит записей (по умолчанию 100, максимум 500)
        in: query
        name: limit
        type: integer
      - description: Смещение
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.UserResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Список пользователей организации
      tags:
      - users
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные пользователя
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Создаёт нового пользователя
      tags:
      - users
  /users/{id}:
    delete:
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удалить пользователя
      tags:
      - users
    get:
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить пользователя по ID
      tags:
      - users
...
securityDefinitions:
  BearerAuth:
    description: 'Введите: Bearer <access_token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"